Hairball now seems to work
This commit represents standalone Hairball reaching a usable state. The
client-side currently has its own 'flavor' which may or may not work
equally well, but this will be eliminated with some refactoring. Note
that this version doesn't actually work with S3 buckets either, but is
able to do async file input, which is a first step to getting there.